[master] refactor to separate tmp lock class
